+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -113,13 +113,13 @@ describe('Datagrid Colspan Frozen Column Tests', () => {
     await utils.checkForErrors();
   });
 
-  it('Should marhup frozen colspan columns', async () => {
-    expect(await element(by.css('#datagrid .left tr:nth-child(3) td:nth-child(2)')).getAttribute('class')).toEqual('is-last-visible l-left-text');
+  it('Should markup frozen colspan columns', async () => {
+    expect(await element(by.css('#datagrid .left tr:nth-child(3) td:nth-child(2)')).getAttribute('class')).toEqual('l-left-text');
     expect(await element(by.css('#datagrid .left tr:nth-child(3) td:nth-child(2)')).getAttribute('colspan')).toEqual('1');
     expect(await element(by.css('#datagrid .center tr:nth-child(3) td:nth-child(3)')).getAttribute('class')).toEqual('');
     expect(await element(by.css('#datagrid .center tr:nth-child(3) td:nth-child(3)')).getAttribute('colspan')).toEqual(null);
 
-    expect(await element(by.css('#datagrid .left tr:nth-child(4) td:nth-child(2)')).getAttribute('class')).toEqual('is-last-visible l-left-text');
+    expect(await element(by.css('#datagrid .left tr:nth-child(4) td:nth-child(2)')).getAttribute('class')).toEqual('l-left-text');
     expect(await element(by.css('#datagrid .left tr:nth-child(4) td:nth-child(2)')).getAttribute('colspan')).toEqual('5');
     expect(await element(by.css('#datagrid .center tr:nth-child(4) td:nth-child(3)')).getAttribute('class')).toEqual('is-hidden l-left-text');
     expect(await element(by.css('#datagrid .center tr:nth-child(4) td:nth-child(3)')).getAttribute('colspan')).toEqual('3');
@@ -1719,6 +1719,25 @@ describe('Datagrid single select tests', () => {
     expect(await element(by.css('#datagrid .datagrid-wrapper tbody tr:nth-child(1)')).getAttribute('class')).toMatch('is-selected');
     expect(await element(by.css('#datagrid .datagrid-wrapper tbody tr:nth-child(2)')).getAttribute('class')).not.toMatch('is-selected');
     expect(await element.all(by.css('#datagrid .datagrid-row.is-selected td:nth-child(1) span')).first().getText()).toEqual('2642205');
+  });
+});
+
+fdescribe('Datagrid spacer row tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datagrid/example-spacer-column');
+
+    const datagridEl = await element(by.css('#datagrid tbody tr:nth-child(1)'));
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should render spacer rows', async () => {
+    expect(await element.all(by.css('#datagrid .datagrid-spacer-column')).count()).toEqual(7);
+    expect(await element.all(by.css('#datagrid .datagrid-header-spacer-column')).count()).toEqual(1);
   });
 });